Safari-609.1.11.1/LayoutTests/fast/parser/hex-entities-length.html

<head>
<title>Hex HTML entities of various lengths</title>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>There should be 6 lines below, each with two copies of the same character.</p>
<ol>
<li>a&#x61;</li>
<li>á¸&#x1e01;</li>
<li>ğ€&#x10400;</li>
<li>ğ‘&#x01044D;</li>
<li>ğ‘&#x001044e;</li>
<li>ğ‘&#x0001044f;</li>
</ol>
</body>
