third_party/blink/web_tests/fast/parser/hex-entities-length-expected.txt

There should be 6 lines below, each with two copies of the same character.

aa
á¸á¸
ğ€ğ€
ğ‘ğ‘
ğ‘ğ‘
ğ‘ğ‘
